re added delta time

+++ include/rotations.h
@@ -13,13 +13,13 @@
 //STRUCTS
 
 struct rotation2D {
-    int h;
-    int v;
+    float h;
+    float v;
 };
 
 //-----------------------------------------------------------------------------------------------//
 //FUNCTIONS
 
-struct rotation2D newRotation2D(int h, int v);
+struct rotation2D newRotation2D(float h, float v);
 
 //-----------------------------------------------------------------------------------------------//
