+++ include/vectors.h
@@ -14,30 +14,30 @@
 
 // may go back to ints, whatever is better later
 struct vector2D {
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 struct vector3D {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
 };
 
 //-----------------------------------------------------------------------------------------------//
 //FUNCTIONS
 
-struct vector2D newVector2D(int x, int y);
+struct vector2D newVector2D(float x, float y);
 
-struct vector3D newVector3D(int x, int y, int z);
+struct vector3D newVector3D(float x, float y, float z);
 
-int manhattanLength2D(struct vector2D v1);
+float manhattanLength2D(struct vector2D v1);
 
-int manhattanLength3D(struct vector3D v1);
+float manhattanLength3D(struct vector3D v1);
 
-int manhattanDistance2D(struct vector2D v1, struct vector2D v2);
+float manhattanDistance2D(struct vector2D v1, struct vector2D v2);
 
-int manhattanDistance3D(struct vector3D v1, struct vector3D v2);
+float manhattanDistance3D(struct vector3D v1, struct vector3D v2);
 
 float euclidianLength2D(struct vector2D v1);
